depraciation function implemented
